Prevent TextView to resize depending on the content to avoid onLayout calls
TextViews got set a new text multiple times a seconde. One of the consequence
is that it calls onLayout on the parent RelativeLayout (StatusOverlayView)
at every change.
Avoid this (as in practice, nothing needs to be relayouted) by setting
TextViews to the parent width, and use gravity to set where the text should be.